feat(frontend): make home page guest welcome experience
