+++ frontend/src/i18n/locales/en/common.json
@@ -12,7 +12,8 @@
     "login": "Login",
     "register": "Register",
     "me": "Profile",
-    "approvals": "Approvals"
+    "approvals": "Approvals",
+    "backendHealth": "Backend health"
   },
   "theme": {
     "toggle": {
@@ -96,6 +97,22 @@
       "submit": "Activate user",
       "submitting": "Activating...",
       "success": "Activated {{username}} ({{role}})."
+    }
+  },
+  "home": {
+    "guest": {
+      "title": "Welcome to VANESSA",
+      "description": "Sign in or create an account to start using your local AI mission console.",
+      "actions": "Guest quick actions",
+      "login": "Go to login",
+      "register": "Create an account"
+    },
+    "authenticated": {
+      "title": "Welcome back",
+      "description": "You are signed in as {{username}}. Open your profile or run backend diagnostics.",
+      "actions": "Signed-in quick actions",
+      "profile": "View profile",
+      "backendHealth": "Run backend diagnostics"
     }
   }
 }
