+++ frontend/src/i18n/locales/es/common.json
@@ -12,7 +12,8 @@
     "login": "Iniciar sesion",
     "register": "Registro",
     "me": "Perfil",
-    "approvals": "Aprobaciones"
+    "approvals": "Aprobaciones",
+    "backendHealth": "Salud del backend"
   },
   "theme": {
     "toggle": {
@@ -96,6 +97,22 @@
       "submit": "Activar usuario",
       "submitting": "Activando...",
       "success": "{{username}} activado ({{role}})."
+    }
+  },
+  "home": {
+    "guest": {
+      "title": "Bienvenido a VANESSA",
+      "description": "Inicia sesion o crea una cuenta para empezar a usar tu consola local de IA.",
+      "actions": "Acciones rapidas para invitados",
+      "login": "Ir a inicio de sesion",
+      "register": "Crear una cuenta"
+    },
+    "authenticated": {
+      "title": "Bienvenido de nuevo",
+      "description": "Has iniciado sesion como {{username}}. Abre tu perfil o ejecuta diagnosticos del backend.",
+      "actions": "Acciones rapidas con sesion iniciada",
+      "profile": "Ver perfil",
+      "backendHealth": "Ejecutar diagnosticos del backend"
     }
   }
 }
